chore/#15: 약관 상수 data 폴더로 분리하기
