+++ src/components/Membership/data/CLAUSE_LIST.tsx
@@ -1,0 +1,9 @@
+interface Event {
+  title: string;
+  content: string;
+}
+
+export const CLAUSE_LIST: Event[] = [
+  { title: '멤버십 등급 기준', content: '' },
+  { title: '유의사항', content: '' },
+];
